Move cl functions around
