--- frontend/args.h
@@ -1,8 +1,0 @@
-#ifndef ARGS_H
-#define ARGS_H
-
-#include "types.h"
-
-extern const form_s Forms[];
-
-#endif /* ARGS_H */
