+++ frontend/forms.h
@@ -1,0 +1,10 @@
+#ifndef CL_FORMS_H
+#define CL_FORMS_H
+
+#include "types.h"
+
+extern const combo_s Combos[];
+extern const ilsuffix_s Il_suffix[];
+extern const form_s Forms[];
+
+#endif
